+++ changelog.md
@@ -5,9 +5,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 
 ## [Unreleased]
-- Functioning GUI
+- Level of difficulty dependent on grade level
 - Login option through MariaDB 
 - Sound effects
+
+## [2.0.0] - 2021-10-29
+
+### Added
+- Comments to the Gui and other small memo's in code 
+- Gui frame for specified exercise + widgets and their locations on the frame
+- Counters displaying the amount of correct and wrong answers
+- Back button that returns to the main menu
+- Display of correct answer when the user inputs the wrong answer
+- Prevention of input besides integers
+- Handlers class
+- return_to_main_menu and check_button_handler methods
+- frm_menu to switch between pages easier
+- limited docstrings
+- counters class
+- ability to press the enter key to input answer
+- Round btn
+- Btn.png
+
+### Changed
+- Positioning of code blocks
+- global amount_wrong, amount_correct variables to class variables
+- Lay-out calculation screen
+- Changed the GUI to match the drawing from discord
+
+### Deprecated
+- check_for_stop method
+- viability_check method
+- viability class
+
+### Removed
+- While loops inside operations methods
+- While loop inside game loop function
+- Output to console for user
+
 
 ## [1.0.1] - 2021-10-01